requirements.txt
selenium==4.15.2
python-telegram-bot==20.8

pip install -r requirements.txt